Fix for Bug #6 Flooding is Not Limited to Water Edge
Flooding could start at any point in the map and not constrained to starting at the water edge. Found that when the startFlood() function was testing if the tile was next to the water edge, it tested if the tile was <= WOODS5 instead of LASTRIVEDGE. Changed so that it would test if tile was <= LASTRIVEDGE and it seemed to fix the bug.