Fix for Bug #6 Underwater Wires Cost Wrong Amount
The cost to build underwater wires was set to $5 instead of $25. Looked at how cost of wires were being determined and saw that there was a layWire function within the RoadLikeTool file. Saw that the cost of the wire was set to $5 even when the tile the wire is built on was underwater. Changed so that the wire cost would be set to under_water_wire cost of $25 and the bug was fixed.