Pirates trigger fires (incomplete)
Calling the pirates disaster now triggers fires in the town. However, they call the fires in the town no matter how far away the pirates are from the city so it looks really odd. Still need to fix so that the fires are only called when/if the pirates get close enough.