Implemented Destruction Feature
Implemented a feature where every so often, the pirate ship will randomly select tiles within a certain range (distance away from ship position) and check if the tile is a building or empty. If it is part of the city, it will set the tile aflame. If it is empty, then it will convert the tile into rubble. Still need to change visuals of ship and add a new sound. Want to hash out tile destruction function so that it focuses more on city part of map.