Completed Pirate Disaster!
Added the Pirate Ship sprites, changed the doPirates() function to call the new sprite, figured out and hashed out the random tile attacks from the pirates, and decreased the time the pirate ship sails around by the amount of seaports there are on the map. The less time the pirate ship sails, the less damage it does so now there is an incentive to build seaports! I also added in two new sounds - a pirate's "arr" cry and the sound of cannons when the pirate ship attacks!